Extract RequestValidator to belgif-rest-problem-validator (#22)
Fixes #16.
