+++ belgif-rest-problem-validator/src/main/java/io/github/belgif/rest/problem/validation/ZeroOrAllOfValidator.java
