+++ belgif-rest-problem-validator/src/test/java/io/github/belgif/rest/problem/validation/RejectedInputValidatorTest.java
